backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "backend"
version = "0.1.0"
description = "Backend for AGENTZ"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "sqlalchemy",
    "crewai==0.11.0",
    "boto3==1.34.14",
    "uvicorn[standard]",
    "loguru",
    "python-dotenv",
    "pydantic",
    "pydantic-settings",
    "httpx",
    "python-multipart",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "supabase>=2.3.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0"
]

[tool.setuptools]
packages = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto" 